requirements.txt
aiohttp==3.9.5
paho-mqtt==1.6.1
